+++ Quackademics/src/components/Flashcards/FlashcardSetCreate.jsx
@@ -1,11 +1,11 @@
 import React, { useState } from "react";
 import { Box, Button, Typography} from "@mui/material";
-import supabase from "../libs/supabaseAdmin";
+import supabase from "../../libs/supabaseAdmin";
 
 //TODO: Make page parse user_id
-const FlashcardSetEdit = ({close, refreshFlashcardSets, flashcardSet}) => {
-    const [setName, setSetName] = useState(flashcardSet.set_name);
-    const [tags, setTags] = useState(flashcardSet.tags);
+const FlashcardSetCreate = ({close, refreshFlashcardSets}) => {
+    const [setName, setSetName] = useState('');
+    const [tags, setTags] = useState('');
     const [errorMessage, setErrorMessage] = useState('');
 
     const handleSetNameEntry = (event) => {
@@ -16,40 +16,46 @@ const FlashcardSetEdit = ({close, refreshFlashcardSets, flashcardSet}) => {
         setTags(event.target.value);
     };
 
-    const updateFlashcardSet = async () => {
+    const createFlashcardSet = async () => {
         if(setName == ''){
             setErrorMessage("Set name cannot be blank.");
             return [];
         }
 
         const { data, error } = await supabase
-            .from("flashcard_set")
-            .update({ set_name: setName, tags: tags })
-            .eq("set_id", flashcardSet.set_id)
+        .from("flashcard_set")
+        .insert([
+            {
+            //Hardcoded user_id for now
+                user_id: 42069,
+                set_name: setName,
+                tags: tags
+            },
+            ])
             .select();
-    
-            if (error) {
-                console.error("Error updating data:", error);
-                if(error.code == 23505){
-                    setErrorMessage("The set name must be unique");
-                }
-                return [];
-            }
-            else{
-                close();
-                if(errorMessage != ''){
-                    setErrorMessage('');
-                }
-            }
 
-            return data;
-        };
-
-        const updateProcess = async () => {
-            await updateFlashcardSet();
-            await refreshFlashcardSets();
-            return;
+        if (error) {
+            console.error("Error inserting data:", error);
+            if(error.code == 23505){
+                setErrorMessage("The set name must be unique");
+            }
+            return [];
         }
+        else{
+            close();
+            if(errorMessage != ''){
+                setErrorMessage('');
+            }
+        }
+
+        return data;
+    };
+
+    const createProcess = async () => {
+        await createFlashcardSet();
+        await refreshFlashcardSets();
+        return;
+    }
 
     return (
         <Box
@@ -71,11 +77,11 @@ const FlashcardSetEdit = ({close, refreshFlashcardSets, flashcardSet}) => {
             </Box>
             <Typography sx={{ color: "red", ml: 1}}> {errorMessage} </Typography>
             <Box sx={{display: "flex", justifyContent: "end", alignItems: "end"}}>
-                <Button title="Save" sx={{backgroundColor: "cornflowerblue", color:"white", mr: 1, mb: 1}} onClick={updateProcess}>Save</Button>
+                <Button title="Save" sx={{backgroundColor: "cornflowerblue", color:"white", mr: 1, mb: 1}} onClick={createProcess}>Save</Button>
                 <Button title="Cancel" sx={{backgroundColor: "cornflowerblue", color:"white", mr: 1, mb: 1}} onClick={close}>Cancel</Button>
             </Box>
         </Box>
     );
 };
 
-export default FlashcardSetEdit;
+export default FlashcardSetCreate;
