+++ Quackademics/src/components/Flashcards/FlashcardSetDelete.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Box, Button, Typography} from "@mui/material";
-import supabase from "../libs/supabaseAdmin";
+import supabase from "../../libs/supabaseAdmin";
 
 //TODO: Make page parse user_id
 const FlashcardSetDelete = ({close, set_id, refreshFlashcardSets}) => {
